requirements.txt
numpy==1.26.2
matplotlib==3.7.1
biopython==1.81
scipy==1.11.4
